requirements.txt
streamlit
zeep
requests
python-dotenv
psycopg2-binary
lxml
cryptography
